Move ui-chunks TS to subdir
