+++ src/index.html
@@ -220,7 +220,7 @@
       </div>
     </main>
 
-    <script type="module" src="js/ui-chunks.ts"></script>
+    <script type="module" src="js/ui-chunks/index.ts"></script>
     <script type="module" src="js/gamemodes.ts"></script>
   </body>
 </html>
