+++ src/js/ui-chunks/replay.ts
@@ -1,61 +1,3 @@
-export class Graph {
-  static popover?: HTMLDivElement;
-
-  static showPopover(el: HTMLSpanElement) {
-    // Not using popover API since anchor stuff doesn't have widespread support + lacking docs
-    this.popover ??= document.querySelector('#RunGraphPopover');
-    this.popover.style.left = `${parseInt(el.style.left.slice(0, -1)) - 10}%`;
-    this.popover.style.top = `${parseInt(el.style.top.slice(0, -1)) + 5}%`;
-    this.popover.querySelector('#Time').textContent = el.dataset['time'];
-    const diff = this.popover.querySelector('#Difference');
-    diff.textContent = el.dataset['diff'];
-    diff.classList.toggle('ahead', el.dataset['ahead'] === 'true');
-
-    this.popover.classList.add('visible');
-  }
-
-  static hidePopover() {
-    this.popover?.classList.remove('visible');
-  }
-
-  static {
-    addEventListener('DOMContentLoaded', () =>
-      document
-        .querySelectorAll('#Graph span.point.interactive')
-        .forEach((point: HTMLSpanElement) => {
-          point.addEventListener('mouseover', () => this.showPopover(point));
-          point.addEventListener('mouseout', () => this.hidePopover());
-        })
-    );
-  }
-}
-
-export class MapSelector {
-  static getRow(rowIdx: number): HTMLDivElement {
-    return document.querySelectorAll<HTMLDivElement>(
-      '#MapSelectorFirstRow, #MapSelectorFirstRow ~ div'
-    )[rowIdx];
-  }
-
-  static toggleFavorite(rowIdx: number): void {
-    const button = this.getRow(rowIdx).querySelector('& > .faves');
-    button.classList.toggle('selected');
-  }
-
-  static fakeDownload(rowIdx: number): void {
-    const button = this.getRow(rowIdx).querySelector('& > .download');
-    const duration = Math.random() * 2 + 0.5;
-    if (button.classList.contains('playable')) return;
-    button.classList.add('downloading');
-    button.querySelector<HTMLSpanElement>('& > span').style.transitionDuration =
-      `${duration}s`;
-    setTimeout(() => {
-      button.classList.remove('downloading');
-      button.classList.add('playable');
-    }, duration * 1000);
-  }
-}
-
 // Absolutely ridiculous overkill but whatever it was FUN
 export class Replay {
   static readonly segmentTimes = [
@@ -185,6 +127,3 @@ export class Replay {
     this.updateProgress();
   }
 }
-
-// Not unreasonable using globals for such a small site.
-Object.assign(window, { MapSelector, Replay });
